+++ src/pages/EditServices/editServiceComponents/TermsCheckbox.jsx
